rearranged the main structure
removed Flask and template folder, moved the tests folder into the old 'app' (overlap_hours) folder, put all of the folders into one big folder, added new .gitignore inside and created new task folder containing the exercise

made main.py the entry point for the program

moved the .gitignore in the overlap folder

fixed import for main logic

added user inputs in the main
